transparencia_servidores_executivo_federal/requirements.txt
requests>=2.31.0
Office365-REST-Python-Client>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
